+++ src/entity/user.rs
@@ -8,8 +8,8 @@ pub struct Model {
     pub id: Uuid,
     #[sea_orm(unique, indexed)]
     pub user_id: i64,
-    pub total_credits: f64,
-    pub credits_remaining: f64,
+    pub total_credits: i64,
+    pub credits_remaining: i64,
     pub subscription_status: bool,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
